+++ data-raw/dev_package.R
@@ -3,7 +3,7 @@ devtools::document()
 
 # add, commit, push
 gert::git_add("*")
-gert::git_commit("version 0.0.0.9089")
+gert::git_commit("version 0.0.0.9090")
 gert::git_push()
 
 pak::pak("blaserlab/blaseRtemplates")