+++ NEWS.md
@@ -144,3 +144,8 @@
 
 * made easy install update to latest version when linking to cache 
 * added load_process template for scrnaseq
+
+# blaseRtemplates 0.0.0.9090
+
+* edited load_process template
+* provided option to use pak for easy_install